Création du new, modify et delete Comment API

+++ Backend/models/comment.js
@@ -7,10 +7,10 @@ exports.newComment = (comment_post, user_id, post_id) => {
     })
 };
 
-exports.modify = (comment_post, comment_id, post_id) => {
+exports.modify = (comment_post, comment_id, user_id, post_id) => {
     try {
         return new Promise((resolve, reject) => {
-            connectionDB.query("UPDATE Comment SET comment_post = ? WHERE id = ? AND post_id = ?;", [comment_post, comment_id, post_id], (error, results) => {
+            connectionDB.query("UPDATE Comment SET comment_post = ?, date_comment = CURRENT_TIMESTAMP WHERE id = ? AND user_id = ? AND post_id = ?;", [comment_post, comment_id, user_id, post_id], (error, results) => {
                 if (results === undefined) {
                     reject ("Erreur dans la requête");
                 }
@@ -25,10 +25,10 @@ exports.modify = (comment_post, comment_id, post_id) => {
     };
 };
 
-exports.delete = (comment_id, post_id) => {
+exports.delete = (comment_id, user_id, post_id) => {
     try {
         return new Promise((resolve, reject) => {
-            connectionDB.query("DELETE FROM Comment WHERE id = ? AND post_id = ?;", [comment_id, post_id], (error, results) => {
+            connectionDB.query("DELETE FROM Comment WHERE id = ? AND user_id = ? AND post_id = ?;", [comment_id, user_id, post_id], (error, results) => {
                 if (results === undefined) {
                     reject ("Erreur dans la requête");
                 }
